feat: file setting of the users route improved
